ruff: PLR0913 the RIGHT way — config/interface extraction, drop dead arg, max-args back to 5
Revert the max-args=7 shortcut; fix all 12 too-many-arg functions cardiac-style:
- ModelCfg (pydantic) + HParams.model_cfg(): ConvVAE/InpaintAE take one cfg, not 6 hparams
- Method(fit, score) interface object: harness.run(name, Method, ...) — 6 run_* callers updated
- FitCfg (patchcore.fit), FpfhCfg (fpfh_for_sample), PP (store.build): group the knobs
- by_defect(ScoreArrays), preprocess(PP): kept from before
- _profiles(shape,...), _score_maps(...,hw,...), export_one(sample,...): pack dims/coords to a tuple
- load_split: DROP dead  arg (0 callers passed it) -> 6 legit query/load params -> bare noqa
  (the one genuine data-accessor case; cardiac uses noqa sparingly the same way)

Backlog 0 at default max-args=5. Models forward-tested, 23 tests green, all runners import.

+++ core/data/dataset.py
@@ -44,8 +44,8 @@ class GpuSplit:
         return self.x.shape[0]
 
 
-def load_split(split=None, label=None, cats=None, channels=("xyz",),
-               device="cuda", dtype=torch.float32, size=None) -> GpuSplit:
+def load_split(split=None, label=None, cats=None, channels=("xyz",),  # noqa: PLR0913
+               device="cuda", size=None) -> GpuSplit:
     """Query the store (filter by split / label / category) -> GPU-resident tensors."""
     df = store.load(size=size or pp.SIZE)
     if cats:
@@ -56,4 +56,4 @@ def load_split(split=None, label=None, cats=None, channels=("xyz",),
         df = df.filter(pl.col("label") == label)
     if len(df) == 0:
         raise ValueError(f"empty split (split={split}, label={label}, cats={cats}) — build the store first")
-    return GpuSplit(df, channels=channels, device=device, dtype=dtype)
+    return GpuSplit(df, channels=channels, device=device)
